requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# PyTorch and transformers
torch>=1.10.0
transformers>=4.18.0
accelerate>=0.6.0
datasets>=2.0.0
sentencepiece>=0.1.96
protobuf>=3.19.0

# API dependencies
fastapi>=0.78.0
uvicorn>=0.17.0
pydantic>=1.9.0
python-multipart>=0.0.5
httpx>=0.23.0

# Data processing
nltk>=3.6.0
spacy>=3.2.0
kaggle>=1.5.12
kagglehub>=0.2.0

# Utilities
tqdm>=4.62.0
PyYAML>=6.0
python-dotenv>=0.20.0
joblib>=1.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Monitoring
prometheus-client>=0.14.0
